scraper-status-updater/requirements.txt
kafka-python==2.0.2
protobuf==3.20.3
pandas==2.1.4
